requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
Django==4.1.2
django-filter==22.1
django-heroku==0.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
Markdown==3.4.1
pycparser==2.21
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
waitress==2.0.0
Werkzeug==2.0.1